update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@299 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ dice/src/parser/Converter.cpp
@@ -245,7 +245,7 @@ long dice::parser::hex_to_long (char *s, IntType& t)
 					t = INT_ULONG;
 					return m;
 				}
-				n = m;
+				n = (long)m;
 				if (n < 0 &&
 					std::numeric_limits<int>::min() <= n &&
 					n <= std::numeric_limits<int>::max())
